@@ -1,44 +1,38 @@
 apply plugin: 'com.android.library'
+apply plugin: 'maven-publish'
 
-ext {
-    bintrayRepo = 'maven'
-    bintrayName = 'android-pdf-viewer'
+task androidSourcesJar(type: Jar) {
+    archiveClassifier.set('sources')
+    from android.sourceSets.main.java.srcDirs
+}
 
-    publishedGroupId = 'com.github.barteksc'
-    libraryName = 'AndroidPdfViewer'
-    artifact = 'android-pdf-viewer'
-
-    libraryDescription = 'Android view for displaying PDFs rendered with PdfiumAndroid'
-
-    siteUrl = 'https://github.com/barteksc/AndroidPdfViewer'
-    gitUrl = 'https://github.com/barteksc/AndroidPdfViewer.git'
-
-    libraryVersion = '3.2.0-beta.1'
-
-    developerId = 'barteksc'
-    developerName = 'Bartosz Schiller'
-    developerEmail = 'barteksch@boo.pl'
-
-    licenseName = 'The Apache Software License, Version 2.0'
-    licenseUrl = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-    allLicenses = ["Apache-2.0"]
+afterEvaluate {
+    publishing {
+        publications {
+            mavenJava(MavenPublication) {
+                groupId = 'com.github.awxkee'
+                artifactId = 'AndroidPdfViewer'
+                version = '1.0.0'
+                from components.release
+                artifact androidSourcesJar
+            }
+        }
+    }
 }
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 33
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 28
+        targetSdkVersion 33
         versionCode 1
-        versionName "3.2.0-beta.1"
+        versionName "1.0.0"
     }
 
 }
 
 dependencies {
-    implementation 'com.android.support:support-compat:28.0.0'
-    api 'com.github.barteksc:pdfium-android:1.9.0'
+    implementation 'androidx.core:core:1.9.0'
+    api 'com.github.awxkee:PdfiumAndroid:1.0.2'
 }
-
-apply from: 'bintray.gradle'
